Code_Plagiarism_Checker/requirements.txt
numpy
scipy
mpmath
matplotlib
sympy
httpx
pillow
pandas
beautifulsoup4
scikit-learn
faiss-cpu
tiktoken
tqdm
tree-sitter
openai
rank-bm25
transformers
seaborn
GitPython
sentence-transformers
groq
pydantic